backend/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.47.2
anyio==4.8.0
attrs==25.1.0
autogen-agentchat==0.4.9.2
autogen-core==0.4.9.2
autogen-ext==0.4.9.2
backoff==2.2.1
beautifulsoup4==4.12.3
browser-use==0.1.40
cachetools==5.5.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
et-xmlfile==1.1.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
gotrue==2.11.3
gradio==5.15.0
gradio_client==1.7.0
greenlet==3.1.1
gspread==6.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
hyperframe==6.1.0
idna==3.6
importlib_metadata==8.6.1
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
langchain-anthropic==0.3.3
langchain-core==0.3.39
langchain-ollama==0.2.2
langchain-openai==0.3.1
langsmith==0.3.11
lxml==5.1.1
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
multidict==6.1.0
numpy==1.26.4
oauthlib==3.2.2
ollama==0.4.7
openai==1.70.0
openpyxl==3.1.2
opentelemetry-api==1.31.0
orjson==3.10.15
packaging==24.2
pandas==2.2.1
pdf2image==1.17.0
pillow==11.1.0
playwright==1.49.1
postgrest==0.19.3
posthog==3.16.0
propcache==0.2.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
PyMuPDF==1.25.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.1
PyYAML==6.0.2
realtime==2.3.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruff==0.9.5
safehttpx==0.1.6
semantic-version==2.10.0
setuptools==75.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
